--- a/arch/arm64/kvm/.gitignore
+++ b/arch/arm64/kvm/.gitignore
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: GPL-2.0-only
+hyp_constants.h