… Documentation/admin-guide/sysctl/kernel.rst
… Documentation/process/programming-language.rst
… Documentation/translations/it_IT/process/programming-language.rst
… Documentation/translations/zh_CN/process/programming-language.rst
… Documentation/translations/zh_TW/process/programming-language.rst
… arch/arm64/include/asm/perf_event.h
… arch/arm64/kernel/perf_callchain.c
… arch/arm64/kernel/vdso32/Makefile
… arch/arm64/kvm/Makefile
… arch/arm64/kvm/arm.c
… arch/arm64/kvm/hyp/Makefile
… arch/arm64/kvm/hyp/hyp-constants.c
… arch/arm64/kvm/hyp/include/hyp/switch.h
… arch/arm64/kvm/hyp/include/nvhe/mm.h
… arch/arm64/kvm/hyp/nvhe/mem_protect.c
… arch/arm64/kvm/hyp/nvhe/mm.c
… arch/arm64/kvm/hyp/nvhe/setup.c
… arch/arm64/kvm/hyp/vhe/switch.c
… arch/arm64/kvm/perf.c
… arch/arm64/kvm/pkvm.c
… arch/arm64/kvm/pmu.c
… arch/csky/kernel/perf_callchain.c
… arch/nds32/kernel/perf_event_cpu.c
… arch/powerpc/perf/core-book3s.c
… arch/riscv/kernel/perf_callchain.c
… arch/s390/include/asm/ctl_reg.h
… arch/s390/include/asm/entry-common.h
… arch/s390/include/asm/lowcore.h
… arch/s390/include/asm/nmi.h
… arch/s390/include/asm/pai.h