**perf support :**

Add support perf script: Add missing fields in usage hint
Add support 'brstackinsnlen' for branch stacks
s390/pai: add support for cryptography counters
Add support for PAI Extension 1 NNPA counters
perf: Rewrite core context handling
perf: Optimize perf_cgroup_switch()
perf evsel amd: Fix IBS error message
perf evsel: Improve AMD IBS (Instruction-Based Sampling) error handling message

* perf: Rewrite core context handling
```javascript
Current design:

    Currently we have a per task and per cpu perf_event_contexts:

      task_struct::perf_events_ctxp[] <-> perf_event_context <-> perf_cpu_context
           ^                                 |    ^     |           ^
           `---------------------------------'    |     `--> pmu ---'
                                                  v           ^
                                             perf_event ------'
  ```
```javascript                                           
  Proposed design:

    New design proposed by this patch reduce to a single task context and
    a single CPU context but adds some intermediate data-structures:

      task_struct::perf_event_ctxp -> perf_event_context <- perf_cpu_context
           ^                           |   ^ ^
           `---------------------------'   | |
                                           | |    perf_cpu_pmu_context <--.
                                           | `----.    ^                  |
                                           |      |    |                  |
                                           |      v    v                  |
                                           | ,--> perf_event_pmu_context  |
                                           | |                            |
                                           | |                            |
                                           v v                            |
                                      perf_event ---> pmu ----------------'

```

**Run Test cases**
```javascript
#cd /kernel/tools/perf/
#sudo make perf
#cp perf /usr/bin/
```
* Add support 'brstackinsnlen' for branch stacks
```javascript
./perf script -F ip,brstackinsn,brstackinsnlen
```

```javascript
root@volcano9b5e-os:/home/amd/ukernel/kernel/tools/perf# ./perf script -F ip,brstackinsn,brstackinsnlen
ffffffff88b42f4b
        amd_pmu_lbr_read+113:
        ffffffff88a0a4d1        insn: e8 ba ad 06 00            ilen: 5 # PRED
        native_read_msr:
        ffffffff88a75290        insn: 53        ilen: 1
        ffffffff88a75291        insn: 89 f9     ilen: 2
        ffffffff88a75293        insn: 0f 32     ilen: 2
        ffffffff88a75295        insn: 48 c1 e2 20       ilen: 4
        ffffffff88a75299        insn: 48 89 d3  ilen: 3
        ffffffff88a7529c        insn: 48 09 c3  ilen: 3
        ffffffff88a7529f        insn: 66 90     ilen: 2
        ffffffff88a752a1        insn: 48 89 d8  ilen: 3
        ffffffff88a752a4        insn: 5b        ilen: 1
        ffffffff88a752a5        insn: c3                        ilen: 1 # PRED
```
* perf/amd/ibs: Make IBS a core pmu
```javascript
./perf record -e cycles:pp ls
[ perf record: Woken up 3 times to write data ]
[ perf record: Captured and wrote 0.024 MB perf.data (11 samples) ]
```
* perf: Optimize perf_cgroup_switch()
```javascript
$ taskset -c 0 ./perf bench sched pipe -l 10000 -T AAA,BBB
```
```javascript
root@volcano9b5e-os:/home/amd/Narasimha/Linux_Backport/tools/perf# taskset -c 0 ./perf bench sched pipe -l 10000 -T AAA,BBB
# Running 'sched/pipe' benchmark:
# Executed 10000 pipe operations between two threads
 
     Total time: 0.029 [sec]
 
       2.982400 usecs/op
         335300 ops/sec
```
```javascript
$ taskset -c 0 ./perf bench sched pipe -l 10000 -l 100
```
```javascript
  root@volcano9b5e-os:/home/amd/Narasimha/Linux_Backport/tools/perf# taskset -c 0 ./perf bench sched pipe -l 10000 -l 100
# Running 'sched/pipe' benchmark:
# Executed 100 pipe operations between two processes
 
     Total time: 0.000 [sec]
 
       6.170000 usecs/op
         162074 ops/sec
```
* perf evsel amd: Fix IBS error message
```javascript
$ ./perf record -e ibs_op//k -C 0
```
```javascript
root@volcano9b5e-os:/home/amd/Narasimha/Linux_Backport/tools/perf# ./perf record -e ibs_op//k -C 0
Error:
AMD IBS doesn't support privilege filtering. Try again without the privilege modifiers (like 'k') at the end.
```
* perf evsel: Improve AMD IBS (Instruction-Based Sampling) error handling message
```javascript
$ ./perf record -e ibs_op// true
```
```javascript
root@volcano9b5e-os:/home/amd/Narasimha/Linux_Backport/tools/perf# ./perf record -e ibs_op// true
[ perf record: Woken up 4 times to write data ]
[ perf record: Captured and wrote 0.021 MB perf.data (11 samples) ]
root@volcano9b5e-os:/home/amd/Narasimha/Linux_Backport/tools/perf# ./perf record -a -e ibs_op// true
[ perf record: Woken up 0 times to write data ]
[ perf record: Captured and wrote 0.896 MB perf.data (3557 samples) ]
```
* perf Order the PMU list to fix warning about unordered pmu_ctx_list

```javascript
amd@debian:~/Narasimha/kernel/tools/perf$ ./perf record -e cycles --call-graph lbr -- taskset -c 3 ./a.out & ./perf stat -e cpu-clock,cs -p 340154
[1] 340158
Error:
cycles:u: PMU Hardware or event type doesn't support branch stack sampling.

 Performance counter stats for process id '340154':

         16,786.77 msec cpu-clock:u               #    0.986 CPUs utilized
                 0      cs:u                      #    0.000 /sec

      17.019565123 seconds time elapsed

```

* perf events test suite
```javascript
git clone https://github.com/deater/perf_event_tests.git

[sudo sysctl kernel.perf_event_paranoid=2]

./fuzzer/perf_fuzzer -t OC -s 100
....
....
Close:  5/5 Successful
Read:   0/0 Successful
Write:  0/0 Successful
Ioctl:  0/0 Successful
```